# api/controllers/comment_controller.go
package controllers

import (
	"encoding/json"
	"fmt"
	"io/ioutil"
	"net/http"
	"strconv"

	"github.com/ahmadnurrizal/final-project-scalable-web-services-with-golang/api/auth"
	"github.com/ahmadnurrizal/final-project-scalable-web-services-with-golang/api/models"
	"github.com/ahmadnurrizal/final-project-scalable-web-services-with-golang/api/utils/formaterror"
	"github.com/gin-gonic/gin"
)

// CreateComment godoc
// @Summary     Create Comment
// @Description Add a new Comment
// @Tags        Comment
// @Accept      json
// @Produce     json
// @Param       CreateComment body models.CreateComment true "Comment Data"
// @Param id path int true "Comment ID"
// @Security ApiKeyAuth
// @Success     200  {object} models.Comment
// @Router      /comments/{id} [post]
func (server *Server) CreateComment(c *gin.Context) {

	//clear previous error if any
	errList = map[string]string{}

	photoID := c.Param("id")
	pid, err := strconv.ParseUint(photoID, 10, 64)
	if err != nil {
		errList["Invalid_request"] = "Invalid Request"
		c.JSON(http.StatusBadRequest, gin.H{
			"status": http.StatusBadRequest,
			"error":  errList,
		})
		return
	}

	uid, err := auth.ExtractTokenID(c.Request)
	if err != nil {
		errList["Unauthorized"] = "Unauthorized"
		c.JSON(http.StatusUnauthorized, gin.H{
			"status": http.StatusUnauthorized,
			"error":  errList,
		})
		return
	}

	// check if the user exist:
	user := models.User{}
	err = server.DB.Debug().Model(models.User{}).Where("id = ?", uid).Take(&user).Error
	if err != nil {
		errList["Unauthorized"] = "Unauthorized"
		c.JSON(http.StatusUnauthorized, gin.H{
			"status": http.StatusUnauthorized,
			"error":  errList,
		})
		return
	}
	// check if the post exist:
	photo := models.Photo{}
	err = server.DB.Debug().Model(models.Photo{}).Where("id = ?", pid).Take(&photo).Error
	if err != nil {
		errList["Unauthorized"] = "Unauthorized"
		c.JSON(http.StatusUnauthorized, gin.H{
			"status": http.StatusUnauthorized,
			"error":  errList,
		})
		return
	}

	body, err := ioutil.ReadAll(c.Request.Body)
	if err != nil {
		errList["Invalid_body"] = "Unable to get request"
		c.JSON(http.StatusUnprocessableEntity, gin.H{
			"status": http.StatusUnprocessableEntity,
			"error":  errList,
		})
		return
	}
	comment := models.Comment{}

	err = json.Unmarshal(body, &comment)
	if err != nil {
		errList["Unmarshal_error"] = "Cannot unmarshal body"
		c.JSON(http.StatusUnprocessableEntity, gin.H{
			"status": http.StatusUnprocessableEntity,
			"error":  errList,
		})
		return
	}

	// enter the userid and the photoid. The comment body is automatically passed
	comment.UserID = uid
	comment.PhotoID = pid

	comment.Prepare()
	errorMessages := comment.Validate()
	if len(errorMessages) > 0 {
		errList = errorMessages
		c.JSON(http.StatusUnprocessableEntity, gin.H{
			"status": http.StatusUnprocessableEntity,
			"error":  errList,
		})
		return
	}

	commentCreated, err := comment.SaveComment(server.DB)
	if err != nil {
		errList := formaterror.FormatError(err.Error())
		c.JSON(http.StatusInternalServerError, gin.H{
			"status": http.StatusInternalServerError,
			"error":  errList,
		})
		return
	}
	c.JSON(http.StatusCreated, gin.H{
		"status":   http.StatusCreated,
		"response": commentCreated,
	})
}

// GetComments godoc
// @Summary Get All Comment
// @Description Retrieve all comment
// @Tags Comment
// @Accept json
// @Produce json
// @Success 200 {array} models.Comment
// @Router /comments [get]
func (server *Server) GetComments(c *gin.Context) {

	comment := models.Comment{}

	comments, err := comment.FindAllComments(server.DB)
	if err != nil {
		errList["No_comment"] = "No Comment Found"
		c.JSON(http.StatusNotFound, gin.H{
			"status": http.StatusNotFound,
			"error":  errList,
		})
		return
	}
	c.JSON(http.StatusOK, gin.H{
		"status":   http.StatusOK,
		"response": comments,
	})
}

// GetCommentByID godoc
// @Summary Get Comment by ID
// @Description Retrieve a comment by ID
// @Tags Comment
// @Accept json
// @Produce json
// @Param id path int true "Comment ID"
// @Success 200 {object} models.Comment
// @Router /comments/{id} [get]
func (server *Server) GetComment(c *gin.Context) {

	commentID := c.Param("id")
	pid, err := strconv.ParseUint(commentID, 10, 64)
	if err != nil {
		errList["Invalid_request"] = "Invalid Request"
		c.JSON(http.StatusBadRequest, gin.H{
			"status": http.StatusBadRequest,
			"error":  errList,
		})
		return
	}
	comment := models.Comment{}

	commentReceived, err := comment.FindCommentByID(server.DB, pid)
	if err != nil {
		errList["No_comment"] = "No Comment Found"
		c.JSON(http.StatusNotFound, gin.H{
			"status": http.StatusNotFound,
			"error":  errList,
		})
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"status":   http.StatusOK,
		"response": commentReceived,
	})
}

// UpdateCommentByID godoc
// @Summary Update Comment by ID
// @Description Update a comment by ID
// @Tags Comment
// @Accept json
// @Produce json
// @Param id path int true "Comment ID"
// @Param UpdateComment body models.UpdateComment true "Comment Data"
// @Security ApiKeyAuth
// @Success 200 {object} models.Comment
// @Router /comments/{id} [put]
func (server *Server) UpdateComment(c *gin.Context) {

	//clear previous error if any
	errList = map[string]string{}

	commentID := c.Param("id")
	// Check if the comment id is valid
	pid, err := strconv.ParseUint(commentID, 10, 64)
	if err != nil {
		errList["Invalid_request"] = "Invalid Request"
		c.JSON(http.StatusBadRequest, gin.H{
			"status": http.StatusBadRequest,
			"error":  errList,
		})
		return
	}
	//CHeck if the auth token is valid and  get the user id from it
	uid, err := auth.ExtractTokenID(c.Request)
	if err != nil {
		errList["Unauthorized"] = "Unauthorized"
		c.JSON(http.StatusUnauthorized, gin.H{
			"status": http.StatusUnauthorized,
			"error":  errList,
		})
		return
	}
	//Check if the comment exist
	origComment := models.Comment{}
	err = server.DB.Debug().Model(models.Comment{}).Where("id = ?", pid).Take(&origComment).Error
	if err != nil {
		errList["No_comment"] = "No Comment Found"
		c.JSON(http.StatusNotFound, gin.H{
			"status": http.StatusNotFound,
			"error":  errList,
		})
		return
	}
	if uid != origComment.UserID {
		errList["Unauthorized"] = "Unauthorized"
		c.JSON(http.StatusUnauthorized, gin.H{
			"status": http.StatusUnauthorized,
			"error":  errList,
		})
		return
	}
	// Read the data commented
	body, err := ioutil.ReadAll(c.Request.Body)
	if err != nil {
		errList["Invalid_body"] = "Unable to get request"
		c.JSON(http.StatusUnprocessableEntity, gin.H{
			"status": http.StatusUnprocessableEntity,
			"error":  errList,
		})
		return
	}
	// Start processing the request data
	comment := models.Comment{}
	err = json.Unmarshal(body, &comment)
	if err != nil {
		errList["Unmarshal_error"] = "Cannot unmarshal body"
		c.JSON(http.StatusUnprocessableEntity, gin.H{
			"status": http.StatusUnprocessableEntity,
			"error":  errList,
		})
		return
	}
	comment.ID = origComment.ID //this is important to tell the model the comment id to update, the other update field are set above
	comment.UserID = origComment.UserID

	comment.Prepare()
	errorMessages := comment.Validate()
	if len(errorMessages) > 0 {
		errList = errorMessages
		c.JSON(http.StatusUnprocessableEntity, gin.H{
			"status": http.StatusUnprocessableEntity,
			"error":  errList,
		})
		return
	}
	commentUpdated, err := comment.UpdateAComment(server.DB)
	if err != nil {
		errList := formaterror.FormatError(err.Error())
		c.JSON(http.StatusInternalServerError, gin.H{
			"status": http.StatusInternalServerError,
			"error":  errList,
		})
		return
	}
	c.JSON(http.StatusOK, gin.H{
		"status":   http.StatusOK,
		"response": commentUpdated,
	})
}

// DeleteCommentByID godoc
// @Summary Delete Comment by ID
// @Description Delete a comment by ID
// @Tags Comment
// @Accept json
// @Produce json
// @Param id path int true "Comment ID"
// @Security ApiKeyAuth
// @Success 200 {string} string "Comment deleted"
// @Router /comments/{id} [delete]
func (server *Server) DeleteComment(c *gin.Context) {

	commentID := c.Param("id")
	// Is a valid comment id given to us?
	pid, err := strconv.ParseUint(commentID, 10, 64)
	if err != nil {
		errList["Invalid_request"] = "Invalid Request"
		c.JSON(http.StatusBadRequest, gin.H{
			"status": http.StatusBadRequest,
			"error":  errList,
		})
		return
	}

	fmt.Println("this is delete comment sir")

	// Is this user authenticated?
	uid, err := auth.ExtractTokenID(c.Request)
	if err != nil {
		errList["Unauthorized"] = "Unauthorized"
		c.JSON(http.StatusUnauthorized, gin.H{
			"status": http.StatusUnauthorized,
			"error":  errList,
		})
		return
	}
	// Check if the comment exist
	comment := models.Comment{}
	err = server.DB.Debug().Model(models.Comment{}).Where("id = ?", pid).Take(&comment).Error
	if err != nil {
		errList["No_comment"] = "No Comment Found"
		c.JSON(http.StatusNotFound, gin.H{
			"status": http.StatusNotFound,
			"error":  errList,
		})
		return
	}
	// Is the authenticated user, the owner of this comment?
	if uid != comment.UserID {
		errList["Unauthorized"] = "Unauthorized"
		c.JSON(http.StatusUnauthorized, gin.H{
			"status": http.StatusUnauthorized,
			"error":  errList,
		})
		return
	}

	_, err = comment.DeleteAComment(server.DB)
	if err != nil {
		errList["Other_error"] = "Please try again later"
		c.JSON(http.StatusInternalServerError, gin.H{
			"status": http.StatusInternalServerError,
			"error":  errList,
		})
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"status":   http.StatusOK,
		"response": "Comment deleted",
	})
}

func (server *Server) GetUserComments(c *gin.Context) {

	userID := c.Param("id")
	// Is a valid user id given to us?
	uid, err := strconv.ParseUint(userID, 10, 64)
	if err != nil {
		errList["Invalid_request"] = "Invalid Request"
		c.JSON(http.StatusBadRequest, gin.H{
			"status": http.StatusBadRequest,
			"error":  errList,
		})
		return
	}
	comment := models.Comment{}
	comments, err := comment.FindUserComments(server.DB, uint32(uid))
	if err != nil {
		errList["No_comment"] = "No Comment Found"
		c.JSON(http.StatusNotFound, gin.H{
			"status": http.StatusNotFound,
			"error":  errList,
		})
		return
	}
	c.JSON(http.StatusOK, gin.H{
		"status":   http.StatusOK,
		"response": comments,
	})
}